fixed port type

+++ cfgparser_core/src/models/core.rs
@@ -1,11 +1,11 @@
 #[derive(serde::Deserialize, Debug)]
 pub struct Configuration {
     pub host: String,
-    pub port: i8,
+    pub port: i64,
 }
 
 impl Configuration {
-    pub fn new(host: String, port: i8) -> Configuration {
+    pub fn new(host: String, port: i64) -> Configuration {
         Configuration {
             host: host,
             port: port,
